Refactor HexBinaryValue and create HexStringFactory
This commit extracts the HexStringFactory class from the HexBinaryValue class to offer
the generation of hex strings without any dependency on the HexBinaryValue class.
